refactor: invalid url handling
