+++ src/app/not-found.tsx
@@ -8,26 +8,8 @@
  *        Summary: Reroutes user to homepage on 404 (invalid URL)
  *
  **************************************************************/
-"use client";
-
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { toast } from "sonner";
+import { redirect } from "next/navigation";
 
 export default function NotFound() {
-    const router = useRouter();
-
-    useEffect(() => {
-        // Timer so toast can show
-        setTimeout(() => {
-            toast.error(
-                "Invalid URL. You have been redirected to the homepage.",
-            );
-        }, 50);
-
-        // Reroute to homepage
-        router.replace("/");
-    }, [router]);
-
-    return null;
+    redirect("/?toast=invalid-url");
 }
